requirements.txt
torch>=1.9.0
transformers>=4.15.0
Pillow>=8.3.2
matplotlib>=3.4.3
numpy>=1.21.2
argparse>=1.4.0 